+++ bubblebbs/postutils.py
@@ -22,6 +22,44 @@ from markdown.extensions.wikilinks import WikiLinkExtension
 from . import config
 
 
+def reference_links(post_model, message: str, reply_to: int = None) -> str:
+    """Parse @id links.
+
+    Explicitly avoids generating links within links.
+
+    """
+
+    def replace(match):
+        at_number = int(match.group(1))
+
+        # Construct the link based on determining if this is
+        # a reference to a thread or if it's a reference to a reply
+        post_referenced = post_model.query.get(at_number)
+        if not post_referenced:
+            valid = False
+        elif post_referenced.reply_to:
+            link = '%d#%d' % (post_referenced.reply_to, post_referenced.id)
+            valid = True
+        else:
+            link = str(post_referenced.id)
+            valid = True
+
+        if valid:
+            return '<a href="/threads/%s" class="reflink">@%d</a>' % (link, at_number)
+        else:
+            return '<span class="reflink reflink-invalid">@%d</span>' % at_number
+
+    soup = BeautifulSoup(message, 'html.parser')
+    at_link_pattern = re.compile(r'@(\d+)')
+    for text_match in soup.find_all(text=True):
+        if re.search(at_link_pattern, text_match) and text_match.parent.name != 'a':
+            new_text = re.sub(at_link_pattern, replace, text_match)
+            text_match.replace_with(new_text)
+
+    return soup.prettify(formatter=None)
+
+
+
 # FIXME: what if passed a name which contains no tripcode?
 def make_tripcode(form_name: str) -> Tuple[str, str]:
     """Create a tripcode from the name field of a post.
